Fix minor static issues like newlines at end of files; remove extra assertions from ion::Matrix; fix the logic of ion::Matrix::Pool to not go outside boundaries; make matrix allocation counters 64 bit to prevent overflow

+++ common/ionlib/inc/ionlib/matrix.h
@@ -210,8 +210,8 @@ namespace ion
 	template <class T>
 	void InitConvolveThreads(uint32_t num_threads, ConvolveTaskData<T>& task_data);
 
-	uint32_t GetMatrixAllocations();
-	uint32_t GetMatrixDeletions();
+	uint64_t GetMatrixAllocations();
+	uint64_t GetMatrixDeletions();
 }; //namespace ion
 
 #endif //ION_MATRIX_H_
